requirements.txt
streamlit==1.11.1
pandas==1.5.3
plotly==5.10.0
seaborn==0.12.1
Pillow==9.2.0
matplotlib==3.5.3
altair==4.2.0
openpyxl==3.0.10
